+++ src/io/mod.rs
@@ -5,4 +5,5 @@ pub mod csv_points;
 pub mod irap;
 pub mod las;
 pub mod tops;
+pub mod vector;
 pub mod xyz;